ViewModel, View 분리 완료
