+++ app/build.gradle
@@ -37,15 +37,15 @@ android {
 }
 
 dependencies {
-
-    implementation 'io.coil-kt:coil:2.2.2'
+    //viewModel
     implementation "androidx.fragment:fragment-ktx:1.5.3"
     implementation "androidx.activity:activity-ktx:1.6.0"
     implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'com.google.android.material:material:1.6.1'
-    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.4.1'
+    //retrofit
+    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.jakewharton.retrofit:retrofit2-kotlinx-serialization-converter:0.8.0'
     implementation(platform("com.squareup.okhttp3:okhttp-bom:4.10.0"))
     implementation("com.squareup.okhttp3:okhttp")
@@ -53,4 +53,6 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    //coil
+    implementation 'io.coil-kt:coil:2.2.2'
 }
